+++ src/main/java/com/unusualmodding/opposing_force/registry/OPEnchantments.java
@@ -16,22 +16,24 @@ import net.minecraftforge.registries.RegistryObject;
 
 @EventBusSubscriber(modid = OpposingForce.MOD_ID, bus = EventBusSubscriber.Bus.MOD)
 public class OPEnchantments {
+
     public static final DeferredRegister<Enchantment> ENCHANTMENTS = DeferredRegister.create(ForgeRegistries.ENCHANTMENTS, OpposingForce.MOD_ID);
 
     public static final EnchantmentCategory TESLA_BOW = EnchantmentCategory.create("tesla_bow", (item -> item == OPItems.TESLA_BOW.get()));
 
-    // Opposing Force categories
-    public static final RegistryObject<Enchantment> BIG_ELECTRIC_BALL = ENCHANTMENTS.register("capacitance", () -> new OPEnchantment("capacitance", Enchantment.Rarity.RARE, TESLA_BOW, 3, 20, EquipmentSlot.MAINHAND));
-    public static final RegistryObject<Enchantment> BOUNCY_ELECTRIC_BALL = ENCHANTMENTS.register("rebound", () -> new OPEnchantment("rebound", Enchantment.Rarity.UNCOMMON, TESLA_BOW, 4, 15, EquipmentSlot.MAINHAND));
-
-    // Vanilla categories
+    public static final RegistryObject<Enchantment> CAPACITANCE = ENCHANTMENTS.register("capacitance", () -> new OPEnchantment("capacitance", Enchantment.Rarity.RARE, TESLA_BOW, 3, 20, EquipmentSlot.MAINHAND));
     public static final RegistryObject<Enchantment> KICKBACK = ENCHANTMENTS.register("kickback", () -> new OPEnchantment("kickback", Enchantment.Rarity.RARE, EnchantmentCategory.CROSSBOW, 3, 10, EquipmentSlot.MAINHAND));
+    public static final RegistryObject<Enchantment> QUASAR = ENCHANTMENTS.register("quasar", () -> new OPEnchantment("quasar", Enchantment.Rarity.VERY_RARE, TESLA_BOW, 1, 30, EquipmentSlot.MAINHAND));
+    public static final RegistryObject<Enchantment> REBOUND = ENCHANTMENTS.register("rebound", () -> new OPEnchantment("rebound", Enchantment.Rarity.UNCOMMON, TESLA_BOW, 4, 15, EquipmentSlot.MAINHAND));
 
     public static boolean areCompatible(OPEnchantment enchantment1, Enchantment enchantment2) {
-        if (enchantment1 == KICKBACK.get() && enchantment2 == Enchantments.QUICK_CHARGE) {
+        if (enchantment1 == KICKBACK.get() && (enchantment2 == Enchantments.QUICK_CHARGE)) {
             return false;
         }
-        if (enchantment1 == BIG_ELECTRIC_BALL.get() && enchantment2 == Enchantments.MULTISHOT) {
+        if (enchantment1 == CAPACITANCE.get() && (enchantment2 == Enchantments.MULTISHOT || enchantment2 == QUASAR.get())) {
+            return false;
+        }
+        if (enchantment1 == QUASAR.get() && (enchantment2 == Enchantments.MULTISHOT || enchantment2 == CAPACITANCE.get() || enchantment2 == Enchantments.QUICK_CHARGE || enchantment2 == REBOUND.get())) {
             return false;
         }
         return true;